Fix card selection bug
The bug was because I did not flip the selected state properly.
I did selected == true instead of selected == false or !selected.
The Card component also stored it's own state which interfered with
the state that the parent (Hand) controlled.
To fix this I changed Card to a functional component.