Implement listening to backend
This commit has not been tested as I don't know how to actually
communicate with the backend.
But I was able to get and process some messages from the backend.
Most of what the backend documentation says has not been tested.
Mostly the later half.

Also I get a problem where when the 2nd player connect
it doesn't want to connect to websockets.
I will need to investigate this later.